+++ example/src/App.svelte
@@ -2,7 +2,7 @@
 	import Typewriter from 'Typewriter.svelte'
 </script>
 
-<Typewriter loop on:done={() => console.log('Hello World!')}>
+<Typewriter on:done={() => console.log('Hello World!')}>
   <p>Lorem & Ipsum</p>
   <p>Dolor & Sit</p>
   <p>Amet & Consectetur</p>